Restructure council member requirements
Turns this longer item about the ideal steering council member into
a list of various activities the contributor may be engaged in that
help the broader community. Tries to balance the desire of indicating
this individual is deeply engaged in the community without requiring
they be involved in all facets.