Add Governance Document (#11)
* Steal governance document from Jupyter

https://raw.githubusercontent.com/jupyter/governance/master/governance.md

https://github.com/jupyter/governance/blob/master/LICENSE.md

* Minor/cosmetic changes Jupyter -> Dask

* Council changes

* Rewrite Institutional Partner Section and add Community Partners

For institutional partners we include the option of direct funding and
services support.  We remove the distinction between contributors and
council members, and place the expectation of 1 FTE.

We create the notion of Community Partners

* swap BDFL and Steering Council in general prose

This isn't intended to have any significant change, but should convey
that we expect the Steering Council to make most decisions in practice.

* Bring partners under one heading

* Apply suggestions from code review

Co-Authored-By: mrocklin <mrocklin@gmail.com>

* Rephrase Q&A as a statement, and simplify

* fix typo

* Remove "by the bdfl" in removal section

* BSD -> BSD-3-clause

Co-Authored-By: mrocklin <mrocklin@gmail.com>

* Add section on voting

* remove need for one external person in NumFOCUS subcommittee

* minor edits to governance document

* Restructure council member requirements

Turns this longer item about the ideal steering council member into
a list of various activities the contributor may be engaged in that
help the broader community. Tries to balance the desire of indicating
this individual is deeply engaged in the community without requiring
they be involved in all facets.

* State 1-to-1 of steering council and owners

* Provide guidance on member removal

* Fill out details of subcommittee operation

* Specify conditions of subcommittee dissolution

* Borrow text from member elections for removal

* style changes

* Handle corrections raised by @pitrou

* Handle @aterrel's comments

* Remove a few more references to 'Main'

* Delegate new BDFL decision to steering council